wip: move syncMatriculas from core to queue
