+++ apps/queue/src/index.ts
@@ -1,5 +1,5 @@
 export { addEmailToConfirmationQueue } from './jobs/confirmationEmail/email.js';
 export { addEnrollmentsToQueue } from './jobs/enrollments/updateEnrollments.js';
 export { addUserEnrollmentsToQueue } from './jobs/enrollments/updateUserEnrollments.js';
-export { addSyncToQueue } from './jobs/matriculas/sync.js';
+export { addSyncToQueue, syncMatriculas } from './jobs/matriculas/sync.js';
 export { addTeachersToQueue } from './jobs/enrollments/updateTeachers.js';
